backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.3.0
python-multipart==0.0.6
pydantic==1.10.7
firebase-admin==6.2.0
requests==2.31.0
python-jose==3.3.0
gunicorn==21.2.0 